+++ CHANGELOG.md
@@ -7,6 +7,13 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
 
+## [11.1.0]
+
+- Adds hikari logs to opentelemetry
+- Fetches core and plugin config from env
+- Open Telemetry configuration is now optional
+- Migrates API calls from supertokens.io to supertokens.com
+
 ## [11.0.5]
 
 - Adds all logs to telemetry which were logged with `io/supertokens/output/Logging.java`